requirements.txt
smartapi-python
pyotp
requests
